End of 2/3/2020 meeting. - Jordan Migneault

+++ src/main/java/frc/robot/ControlPanel.java
@@ -60,7 +60,7 @@ public class ControlPanel
         }
     }
 
-    public void Spin(int spinSpeed)
+    public void Spin(double spinSpeed)
     {
         spinner.set(ControlMode.PercentOutput, spinSpeed);
     }
